pyproject.toml
[project]
name = "eko"
version = "0.1.0"
description = ""
authors = [
    {name = "Elton Costa",email = "89019551+ryzxxn@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "chromadb (>=0.6.3,<0.7.0)",
    "fastembed (>=0.6.0,<0.7.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "openai (>=1.69.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
